# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "An API for front-end",
        "title": "DDPlanet Server API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
        "/support/request": {
            "post": {
                "description": "Create new request in DB for further checking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create new request to support",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.CreateSupportRequestInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CreateSupportRequestOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/support/requests": {
            "get": {
                "description": "Get list of support requests according to filter",
                "produces": [
                    "application/json"
                ],
                "summary": "Get list of support requests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit of requests",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "status to filer in",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "unix timestamp time of creation",
                        "name": "tsCreatedFrom",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "unix timestamp time of creation",
                        "name": "tsCreatedTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by phone or ID",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/supporting.SupportRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpext.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.CreateSupportRequestInput": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "secondName": {
                    "type": "string"
                }
            }
        },
        "app.CreateSupportRequestOutput": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "integer"
                }
            }
        },
        "httpext.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "supporting.SupportRequest": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                },
                "closedTs": {
                    "type": "integer"
                },
                "createdTs": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "secondName": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}